requirements.txt
backcall==0.2.0
decorator==4.4.2
Django==2.2
django-debug-toolbar==2.2
django-extensions==3.0.3
django-filter==2.3.0
django-tables2==2.3.1
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.10.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pydotplus==2.0.2
Pygments==2.6.1
pyparsing==2.4.7
pytz==2020.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
